Add some .cc files to put implementations
Header-only libraries are not desired because of increased compilation
time and hindered readability.

SSE is required anyway: otherwise compilation fails already.
