--- src/qx/dummy.cc
@@ -1,4 +1,0 @@
-// NOTE: QX is currently still header-only... so there are no sources.
-
-// If ever it is turned into a normal library, this is where the files should
-// go.
